requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Flask==2.1.2
Flask-Docs==0.6.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
greenlet==1.1.2
hashids==1.3.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mistune==2.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-ordering==0.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
